+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Nextcloud v30 is no longer supported
 - delete extra permissions on file delete
 - replaced general save success message with dedicated messages for each settings section
+- show warning popup when jwt secret is empty and success message otherwise
 
 ## Added
 - plugin description and useful links in admin settings